Move /models/shared to /shared, clean up server logs
