+++ src/server/services/getTokenService.ts
@@ -1,5 +1,5 @@
-import { SHARED_GetTokenAPIResponse } from '@/models/shared'
 import { GH_OAuth_ExchangeClientCodeForAccessTokenAPI } from '@/server/lib/gh-oauth/access_token'
+import { SHARED_GetTokenAPIResponse } from '@/shared/models'
 
 export const exchangeCodeForAccessToken = async (code: string): Promise<SHARED_GetTokenAPIResponse> => {
     const { success, error, access_token } = await GH_OAuth_ExchangeClientCodeForAccessTokenAPI(code)
